+++ CHANGELOG.md
@@ -13,12 +13,13 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Validation of `<div>` elements (e.g., div soup).
 - Validation of `<a>` (link) elements, including generic link text, proper attributes.
-- Added `<button>` text content check
+- Added `<button>` text content check.
 
 **HTML Validation:**
 
 - Validation of `<section>` elements (e.g., appropriate labels).
-- Added `<button>` text content check
+- Added `<button>` text content check.
+- Added `<img>` unique alt check.
 
 ## [0.0.5]- 2024-08-27